drivers/intel/mipi_camera: Change type for gpio_num to uint16_t
gpio_num is used to indicate the GPIO which is taken from gpio_soc_defs.h file.
Support for dynamic generation of ASL file for Camera was added for JSL
when there were less than 256 GPIOs. ADL now has more GPIOs and therefore
uint8_t is not enough any more

Reviewed-on: https://review.coreboot.org/c/coreboot/+/55538
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/drivers/intel/mipi_camera/chip.h
@@ -117,7 +117,7 @@ struct clk_config {
 };
 
 struct gpio_config {
-	uint8_t gpio_num;
+	uint16_t gpio_num;
 };
 
 struct clock_ctrl_panel {
